front: create study view folder
