+++ front/src/applications/operationalStudies/views/Study/components/DateBox.tsx
