+++ front/src/applications/operationalStudies/views/Study/components/ScenarioCard.tsx
