+++ front/src/applications/operationalStudies/views/Study/index.ts
@@ -1,0 +1,3 @@
+import StudyView from './StudyView';
+
+export default StudyView;
