+++ front/src/applications/operationalStudies/views/Study/styles/_scenarioCard.scss
