+++ front/src/applications/operationalStudies/views/Study/styles/_study.scss
@@ -1,3 +1,5 @@
+@use 'scenarioCard';
+
 .study-view {
   .study-details {
     width: 100%;
