+++ front/src/modules/scenario/styles/scenario.scss
@@ -1,4 +1,3 @@
 @use './addOrEditScenarioModal';
 @use './scenarioExplorer';
-@use './scenarioCard';
 @use './scenarioLoaderMessage';
