+++ front/src/styles/scss/applications/_operationalStudies.scss
@@ -1,7 +1,6 @@
 @use 'operationalStudies/home';
 @use 'operationalStudies/nge';
 @use 'operationalStudies/project';
-@use 'operationalStudies/study';
 @use 'operationalStudies/scenario';
 @use 'operationalStudies/manageTimetableItem';
 @use 'operationalStudies/importTimetableItem';
