+++ front/src/styles/styles.scss
@@ -1,3 +1,5 @@
+@use '../applications/operationalStudies/views/Study/styles/study';
+
 @use '../modules/modules.scss';
 @use '../common/common.scss';
